+++ src/components/Card/index.ts
@@ -1,0 +1,2 @@
+export * from './Card';
+export { default } from './Card';
